frequency recovery bit width results

+++ src/+freq_recovery/differential_fxp.m
@@ -152,7 +152,9 @@ function [y, frequency_offset] = differential_fxp(x, training, Rs, ~, T, data_ai
             theta = double(theta_fi) * max_freq;
             y(i, p) = cast(x_float(i, p) * exp(1j * theta), 'like', T.x);
 
-            theta_next = double(theta_fi + delta_theta);
+            % Add in double so the explicit phase wrap below is not
+            % pre-empted by fi-overflow on theta_fi + delta_theta.
+            theta_next = double(theta_fi) + double(delta_theta);
             theta_next = mod(theta_next + theta_wrap, 2 * theta_wrap) - theta_wrap;
             theta_fi = cast(theta_next, 'like', T.theta);
         end
